+++ app/main.py
@@ -1,8 +1,13 @@
 from fastapi import FastAPI
 from fastapi.responses import HTMLResponse
 import asyncio
+import datetime
+
+from app.classes_map import *
 
 app = FastAPI()
+
+WORLD_MAP = World_map()
 
 # ===== PAGE =====================================
 
@@ -16,6 +21,7 @@ async def root():
             <body>
                 <h1>Trains 2024</h1>
                 Welcome to the Trains_2024 project server.
+                <p><a href="docs#">API Documentation</a></p>
             </body>
         </html>
     """
@@ -23,10 +29,60 @@ async def root():
 
 # ===== REQUESTS =====================================
 
-@app.get("/get-map")
-async def get_map():
-    print("========== get num ===============")
-    return {"tiles": [1, 2, 3]}
+get_world_examples = {
+    200: {
+        "description": "Success",
+        "content": {
+            "application/json": {
+                "example": {
+                    "tiles":[{
+                        "id":541,
+                        "coord":[29,18],
+                        "tracks":[],
+                        "type":"grass"
+                    }],
+                    "trains": [],
+                    "server_time":"2024-08-05T12:59:15.272432"
+                }
+            }
+        },
+    },
+}
+
+@app.get("/get-world", responses=get_world_examples)
+async def get_world():
+    """Prepare and provide World's data for API response."""
+    response = {
+        "tiles": WORLD_MAP.prepare_tiles_data_for_response(),
+        "trains": WORLD_MAP.prepare_trains_data_for_response(),
+        "server_time": datetime.datetime.now()
+    }
+    return response
+
+get_trains_examples = {
+    200: {
+        "description": "Success",
+        "content": {
+            "application/json": {
+                "example": {
+                    "trains":[{
+                        "id":541,
+                    }],
+                    "server_time":"2024-08-05T12:59:15.272432"
+                }
+            }
+        },
+    },
+}
+
+@app.get("/get-trains", responses=get_trains_examples)
+async def get_trains():
+    """Prepare and provide train's data for API response."""
+    response = {
+        "trains": WORLD_MAP.prepare_trains_data_for_response(),
+        "server_time": datetime.datetime.now()
+    }
+    return response
 
 # ===== LOOP =====================================